requirements.txt
#python==3.11.13

numpy==1.26.4
pandas==2.2.3

matplotlib==3.7.2
seaborn==0.12.2

scipy==1.15.3
scikit-image==0.25.2

scikit-learn==1.2.2

torch==2.6.0+cu124

Pillow==11.3.0
opencv-python==4.12.0
nibabel==5.3.2

kagglehub==0.3.13

pip install pylibjpeg pylibjpeg-libjpeg

dicom 